Move to be deprecated APIs to their own file
